胡椒

+++ Arduino/ATBicycle/Calc.ino
@@ -1,6 +1,6 @@
 double SPEED(double convert) {
   if(convert == 0) return -1;
-  return (1000 / ((convert * 1000000 / 3600) / tire_size));
+  return (double)((double)1000 / (((double)convert * (double)1000000 / (double)3600) / (double)tire_size));
 }
 
 double RPMtoInterval(double convert) {
@@ -11,4 +11,5 @@ int getRPM() {
   // r/minにする ... convert * 1000 -> 秒に * 60 -> 分に
   return ((double)1000 / (double)cadence * (double)60);
 }
+
